requirements.txt
Django==2.1.7
django-smart-selects==1.5.3
django-widget-tweaks==1.4.1
django-geoposition-2==0.3.6
django-star-ratings==0.7.0
django-filter==2.0.0
djangorestframework==3.9.4
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.7
django-braces==1.13.0
django-crontab==0.7.1
django-filter==2.0.0
django-geoposition-2==0.3.6
django-import-export==1.2.0
django-inline-actions==2.2.0
django-jet==1.0.8
django-model-utils==3.1.2
django-mysql==3.2.0
django-rest-auth==0.9.5
django-star-ratings==0.7.0
django-uuidfield==0.5.0
django-widget-tweaks==1.4.1
djangorestframework==3.9.4
et-xmlfile==1.0.1
fcm-django==0.2.21
idna==2.8
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.0.0
protobuf==3.9.0
psycopg2-binary==2.8.3
pyfcm==1.4.5
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-toolbelt==0.9.1
six==1.12.0
sqlparse==0.3.0
sseclient==0.0.24
swapper==1.1.0
tablib==0.13.0
urllib3==1.25.3
whitenoise==4.1.3
xlrd==1.2.0
xlwt==1.3.0
